chore: refactor project structure and add utility types
- Removed unused image file .
- Introduced new TypeScript utility types in  for better type handling.
- Updated import paths in  and  to use the new  directory structure.
- Added common error handling components in .
- Refactored invoice-related utility functions into .
- Implemented rate limiting functionality in  and Redis connection in .
- Cleaned up unused utility functions in .
