+++ src/app/components/invoice-form/common.tsx
@@ -1,0 +1,9 @@
+import { AlertTriangle } from "lucide-react";
+
+export const ErrorMessage = ({ children }: { children: React.ReactNode }) => {
+  return <p className="mt-1 text-xs text-red-600">{children}</p>;
+};
+
+export const AlertIcon = () => {
+  return <AlertTriangle className="mr-1 inline-block h-3 w-3 text-amber-500" />;
+};
